+++ lib/types/date.dart
@@ -1,3 +1,4 @@
+import 'package:jar/jar.dart';
 import 'package:jar/schema.dart';
 
 class JarDate extends JarSchema<DateTime, JarDate> {
@@ -52,6 +53,15 @@ class JarDate extends JarSchema<DateTime, JarDate> {
       return value.isAtSameMomentAs(otherValue!)
           ? null
           : (message ?? 'Must be equal to $field');
+    });
+  }
+
+  JarDate custom(String? Function(DateTime? value) validator,
+      [String? message]) {
+    return addValidator((value) {
+      if (value == null) return null;
+      final validationResult = validator(value);
+      return validationResult ?? null;
     });
   }
 }